Working multiple powerups display - When collected more than one of bomb count or range powerups, will now display the amount gathered. If reached the max amount allowed, will instead display "max". - Misc changes to Text class in order to accomadate for the new InfoPanel additions. - Moved Text texture setting to Animations instead of InfoPanel to avoid any non-loaded texture issues. - Made game static members private and added getters to access them. Classes only need the data, not to be able to modify them. - Went through the project and tried to remove as many compiler warnings as I could. Still getting fallthrough warnings in Enemy though, as well as a linker library conflict. - New constans of course because why would there not be??

+++ Bomberman/Entity.cpp
@@ -20,8 +20,10 @@ int Entity::getY() const							{ return tileY; }
 
 void Entity::move(float x, float y)					{ sprite.move(sf::Vector2f(x, y)); }
 void Entity::setScale(float x, float y)				{ sprite.setScale(sf::Vector2f(x, y)); }
-void Entity::setPosition(int x, int y)				{ sprite.setPosition(sf::Vector2f(x * _scaledTile + _halfScaled, y * _scaledTile + _halfScaled));
-														tileX = x; tileY = y; }
+void Entity::setPosition(int x, int y)				{ sprite.setPosition(sf::Vector2f
+														(static_cast<float>(x * _scaledTile + _halfScaled),
+														 static_cast<float>(y * _scaledTile + _halfScaled)));
+														 tileX = x; tileY = y; }
 void Entity::setOrigin(float x, float y)			{ sprite.setOrigin(sf::Vector2f(x, y)); }
 void Entity::setTexture(int x, int y)				{ sprite.setTextureRect(sf::IntRect(sf::Vector2i(x, y), _tile)); }
 void Entity::draw(sf::RenderTarget& target,
@@ -46,10 +48,14 @@ bool Entity::intersects(const Entity& other) const
 // I have to create the rectangle to make the comparison myself.
 bool Entity::intersects(int x, int y) const
 {
-	return this->getSprite().getGlobalBounds().					// At position
-		findIntersection(sf::FloatRect							// (tileX * game tile scale,
-		(sf::Vector2f(x * _scaledTile, y * _scaledTile),		// tileY * game tile scale),
-		 sf::Vector2f(_scaledTile, _scaledTile))).has_value();	// with size of game tile scale
+	return this->getSprite().getGlobalBounds().findIntersection
+		(sf::FloatRect
+		(sf::Vector2f											// At position
+			(static_cast<float>(x * _scaledTile),				// (tileX * game tile scale,
+			 static_cast<float>(y * _scaledTile)),				// tileY * game tile scale),
+		sf::Vector2f
+			(static_cast<float>(_scaledTile),					// with size of game tile scale
+			 static_cast<float>(_scaledTile)))).has_value();
 }
